# core-common-sca-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-sca-mgmt
  description: Strong Customer Authentication Management Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/sca
    description: Development Environment
  - url: /
    description: Local Development Environment
paths:
...
  /api/v1/sca/operations/{operationId}/challenges/{challengeId}/attempts/{attemptId}:
    get:
      tags:
        - sca-attempt-controller
      operationId: getAttempt
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAAttemptDTO'
    put:
      tags:
        - sca-attempt-controller
      operationId: updateAttempt
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SCAAttemptDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAAttemptDTO'
    delete:
      tags:
        - sca-attempt-controller
      operationId: deleteAttempt
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /api/sca/v1/operations/{operationId}/audit/{auditId}:
    get:
      tags:
        - sca-audit-controller
      operationId: getAuditEvent
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: auditId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAAuditDTO'
    put:
      tags:
        - sca-audit-controller
      operationId: updateAuditEvent
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: auditId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SCAAuditDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAAuditDTO'
    delete:
      tags:
        - sca-audit-controller
      operationId: deleteAuditEvent
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: auditId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/sca/operations/{operationId}/challenges/{challengeId}/attempts:
    get:
      tags:
        - sca-attempt-controller
      operationId: getAllAttempts
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseSCAAttemptDTO'
    post:
      tags:
        - sca-attempt-controller
      operationId: createAttempt
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: challengeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SCAAttemptDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAAttemptDTO'
  /api/sca/v1/operations/{operationId}/audit:
    get:
      tags:
        - sca-audit-controller
      operationId: getAllAudit
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseSCAAuditDTO'
    post:
      tags:
        - sca-audit-controller
      operationId: createAuditEvent
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SCAAuditDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SCAAuditDTO'
components:
  schemas:
...
    SCAAttemptDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        scaChallengeId:
          type: string
          format: uuid
        attemptValue:
          type: string
        attemptedAt:
          type: string
          format: date-time
        success:
          type: boolean
        ipAddress:
          type: string
    SCAAuditDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        scaOperationId:
          type: string
          format: uuid
        scaChallengeId:
          type: string
          format: uuid
        partyId:
          type: string
        eventType:
          type: string
          enum:
            - CREATED
            - ATTEMPTED
            - VERIFIED
            - FAILED
            - EXPIRED
            - CANCELLED
        eventTime:
          type: string
          format: date-time
        details:
          type: string
...
    PaginationResponseSCAAttemptDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/SCAAttemptDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
    PaginationResponseSCAAuditDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/SCAAuditDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.